Change BufferedReader with CSVReader
