+++ src/main/java/com/motorph/AttendanceDataFile.java
@@ -1,18 +1,20 @@
 package com.motorph;
 
-import java.io.BufferedReader;
+import com.opencsv.CSVParser;
+import com.opencsv.CSVParserBuilder;
+import com.opencsv.CSVReader;
+import com.opencsv.CSVReaderBuilder;
+
 import java.io.FileReader;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 
-// Not used because it is an additional feature NOT in the console app
-
-public class AttendanceDataFromFile {
+public class AttendanceDataFile {
     private final ArrayList<Attendance> attendance = new ArrayList<>();
 
-    public AttendanceDataFromFile() {
+    public AttendanceDataFile() {
         String attendanceData = "src/main/resources/Attendance Record.csv";
         dataFromFile(attendanceData);
     }
@@ -22,12 +24,16 @@ public class AttendanceDataFromFile {
 
     // Method that reads the CSV file of attendance data
     public void dataFromFile(String attendanceData) {
-        try (BufferedReader reader = new BufferedReader(new FileReader(attendanceData))){
-            reader.readLine(); // Skip header
-            String line;
+        CSVParser parser = new CSVParserBuilder()
+                .withSeparator(',').build();
+
+        try (CSVReader reader = new CSVReaderBuilder(new FileReader(attendanceData))
+                .withCSVParser(parser).build()){
+            reader.readNext(); // Skip header
+            String[] line;
 
             // Read the csv file line by line
-            while ((line = reader.readLine()) != null) {
+            while ((line = reader.readNext()) != null) {
                     attendance.add(createAttendance(line)); // Add attendance objects to attendance ArrayList
             }
 
@@ -37,14 +43,13 @@ public class AttendanceDataFromFile {
     }
 
     // Converts single lines of CSV data into an attendance Objects
-    private Attendance createAttendance (String lines) {
+    private Attendance createAttendance (String[] values) {
         // Instantiate employee-related class
         Attendance attendance = new Attendance();
         Person person = new Person();
 
         // Parse and set all attendance attributes
         try {
-            String[] values = lines.split(",");
             attendance.getEmployee().setEmployeeNumber(Integer.parseInt(values[0].trim()));
 
             attendance.getEmployee().setPerson(person);
