+++ src/main/java/com/motorph/NewEmployee.form
@@ -3,7 +3,7 @@
   <grid id="27dc6" binding="pnlNewEmployee" layout-manager="GridLayoutManager" row-count="2" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="20" left="20" bottom="20" right="20"/>
     <constraints>
-      <xy x="20" y="20" width="839" height="660"/>
+      <xy x="20" y="20" width="839" height="688"/>
     </constraints>
     <properties/>
     <border type="none"/>
@@ -466,7 +466,7 @@
               </component>
             </children>
           </grid>
-          <grid id="819c8" layout-manager="GridLayoutManager" row-count="1" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+          <grid id="819c8" layout-manager="GridLayoutManager" row-count="2" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
             <margin top="10" left="10" bottom="10" right="10"/>
             <constraints>
               <grid row="3" column="1" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
@@ -481,6 +481,15 @@
                 <properties>
                   <font name="JetBrains Mono" size="14" style="0"/>
                   <text value="Save Employee"/>
+                </properties>
+              </component>
+              <component id="28e2f" class="javax.swing.JButton" binding="btnBack">
+                <constraints>
+                  <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
+                </constraints>
+                <properties>
+                  <font name="JetBrains Mono" size="14" style="0"/>
+                  <text value="Back"/>
                 </properties>
               </component>
             </children>
